feat(agents): enhance agent execution with query-text and resilient outputs
- Add query_text support to BaseAgent and Supervisor for better RAG grounding
- Refactor Supervisor to provide detailed failure reasons for individual agents
- Update AgentEvaluationResult to explicitly track success and error messages

+++ server/modules/agents/contracts.py
@@ -30,7 +30,7 @@ class AgentEvaluationResult:
     agent_name: str
     evaluation_id: uuid.UUID
     document_id: uuid.UUID
-    subtotal: int
+    subtotal: float
     criterion_scores: tuple[CriterionScore, ...]
     summary: str
     model_name: str
